JAVAOS-67:update test case about getObject
Former-commit-id: 5a2636cbebb0356a1523615a14c523a1f228e774

+++ test/JDBCAppendNewTest.java
@@ -836,7 +836,7 @@ public class JDBCAppendNewTest {
         ps.executeUpdate();
         ResultSet rs = ps.executeQuery("select * from pt");
         rs.next();
-        org.junit.Assert.assertEquals("0.00000000",rs.getObject("dataType").toString());
+        org.junit.Assert.assertEquals("0.00000000",((BigDecimal)rs.getObject("dataType")).toString());
     }
 
     @Test
